requirements.txt
fastapi[all]~=0.115.3
pydantic~=2.9.2
uvicorn~=0.32.0
fastapi-restful[all]
tortoise-orm[asyncpg]
dependency-injector>=4.0,<5.0
passlib~=1.7.4
slowapi~=0.1.9
pydantic-settings~=2.6.0
aiohttp~=3.10.10
aioredis~=2.0.1
bcrypt~=4.2.0
PyJWT~=2.9.0
starlette~=0.41.2
redis~=5.2.0
aioboto3~=13.2.0